great_expectations/requirements.txt
# Core ETL dependencies
pandas>=1.5.0
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0

# Great Expectations
great-expectations>=0.18.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Utilities
python-dotenv>=0.19.0
pydantic>=1.10.0
Jinja2==3.0.3

# Development
black>=22.0.0
flake8>=5.0.0
mypy>=0.991
